Enum leptos_router::StaticMode

pub enum StaticMode {
    Upfront,
    Incremental,
}

The mode to use when rendering the route statically. On mode Upfront, the route will be built with the server is started using the provided static data. On mode Incremental, the route will be built on the first request to it and then cached and returned statically for subsequent requests.

Variants

Upfront

Incremental