Function leptos_router::A

pub fn A<H>(cx: Scope, props: AProps<H>) -> impl IntoViewwhere
    H: ToHref + 'static,

An HTML a progressively enhanced to use client-side routing. Client-side routing also works with ordinary HTML <a> tags, but <A> does two additional things:

Correctly resolves relative nested routes. Relative routing with ordinary <a> tags can be tricky. For example, if you have a route like /post/:id, <A href="1"> will generate the correct relative route, but <a href="1"> likely will not (depending on where it appears in your view.)

Sets the aria-current attribute if this link is the active link (i.e., it’s a link to the page you’re on). This is helpful for accessibility and for styling. For example, maybe you want to set the link a different color if it’s a link to the page you’re currently on.

Required Props

cx: Scope

href: [H]

Used to calculate the link’s href attribute. Will be resolved relative to the current route.

children: Children

The nodes or elements to be shown inside the link.

Optional Props

exact: bool

If true, the link is marked active when the location matches exactly; if false, link is marked active if the current route starts with it.

state: State

An object of any type that will be pushed to router state

replace: bool

If true, the link will not add to the browser’s history (so, pressing Back will skip this page.)

class: impl Into<AttributeValue>

Sets the class attribute on the underlying <a> tag, making it easier to style.

id: impl Into<String>

Sets the id attribute on the underlying <a> tag, making it easier to target.