Function leptos_router::ActionForm

pub fn ActionForm<I, O>(props: ActionFormProps<I, O>) -> impl IntoViewwhere
    I: Clone + ServerFn + 'static,
    O: Clone + Serialize + DeserializeOwned + 'static,

Automatically turns a server Action into an HTML form progressively enhanced to use client-side routing.

Encoding

Note: <ActionForm/> only works with server functions that use the default Url encoding. This is to ensure that <ActionForm/> works correctly both before and after WASM has loaded.

Required Props

action: Action<I, Result<O, ServerFnError>>

The action from which to build the form. This should include a URL, which can be generated by default using create_server_action or added manually using leptos_server::Action::using_server_fn.

children: Children

Component children; should include the HTML of the form elements.

Optional Props

class: impl Into<AttributeValue>

Sets the class attribute on the underlying <form> tag, making it easier to style.

error: RwSignal<Option<Box<dyn Error>>>

A signal that will be set if the form submission ends in an error.

node_ref: NodeRef<html::Form>

A NodeRef in which the <form> element should be stored.

noscroll: bool

Sets whether the page should be scrolled to the top when navigating.

attributes: [impl Into<Vec<(&'static str, Attribute)>>](Vec<(&’static str, Attribute)>)

Arbitrary attributes to add to the <form>