Function leptos_router::AnimatedOutlet

pub fn AnimatedOutlet(props: AnimatedOutletProps) -> impl IntoView

Displays the child route nested in a parent route, allowing you to control exactly where that child route is displayed. Renders nothing if there is no nested child.

Animations

The router uses CSS classes for animations, and transitions to the next specified class in order when the animationend event fires. Each property takes a &'static str that can contain a class or classes to be added at certain points. These CSS classes must have associated animations.

outro: added when route is being unmounted

start: added when route is first created

intro: added after start has completed (if defined), and the route is being mounted

finally: added after the intro animation is complete

Each of these properties is optional, and the router will transition to the next correct state whenever an animationend event fires.

Optional Props

class: impl Into<TextProp>

Base classes to be applied to the <div> wrapping the outlet during any animation state.

outro: [&'static str]

CSS class added when route is being unmounted

outro_back: [&'static str]

CSS class added when route is being unmounted, in a “back” navigation

start: [&'static str]

CSS class added when route is first created

intro: [&'static str]

CSS class added while the route is being mounted

intro_back: [&'static str]

CSS class added while the route is being mounted, in a “back” navigation

finally: [&'static str]

CSS class added after other animations have completed.