Function leptos_router::MultiActionForm

pub fn MultiActionForm<I, O>(
    cx: Scope,
    props: MultiActionFormProps<I, O>
) -> impl IntoViewwhere
    I: Clone + ServerFn + 'static,
    O: Clone + Serializable + 'static,

Automatically turns a server MultiAction into an HTML form progressively enhanced to use client-side routing.

Required Props

cx: Scope

action: MultiAction<I, Result<O, ServerFnError>>

The action from which to build the form. This should include a URL, which can be generated by default using create_server_action or added manually using leptos_server::Action::using_server_fn.

children: Children

Component children; should include the HTML of the form elements.

Optional Props

class: impl Into<AttributeValue>

Sets the class attribute on the underlying <form> tag, making it easier to style.

error: RwSignal<Option<Box<dyn Error>>>

A signal that will be set if the form submission ends in an error.

node_ref: NodeRef<html::Form>

A NodeRef in which the <form> element should be stored.

attributes: impl Into<MaybeSignal<AdditionalAttributes>>

Arbitrary attributes to add to the <form>