Function leptos_router::ProtectedRoute

pub fn ProtectedRoute<P, E, F, C>(
    cx: Scope,
    props: ProtectedRouteProps<P, E, F, C>
) -> impl IntoViewwhere
    E: IntoView,
    F: Fn(Scope) -> E + 'static,
    P: Display + 'static,
    C: Fn(Scope) -> bool + 'static,

Describes a route that is guarded by a certain condition. This works the same way as <Route/>, except that if the condition function evaluates to false, it redirects to redirect_path instead of displaying its view.

Required Props

cx: Scope

path: [P]

The path fragment that this route should match. This can be static (users), include a parameter (:id) or an optional parameter (:id?), or match a wildcard (user/*any).

redirect_path: [P]

The path that will be redirected to if the condition is false.

condition: [C]

Condition function that returns a boolean.

view: [F]

View that will be exposed if the condition is true.

methods: [&'static [Method]]

The HTTP methods that this route can handle (defaults to only GET).

Optional Props

ssr: SsrMode

The mode that this route prefers during server-side rendering. Defaults to out-of-order streaming.

children: Children

children may be empty or include nested routes.