Function leptos_router::StaticRoute

pub fn StaticRoute<E, F, P, S>(
    props: StaticRouteProps<E, F, P, S>
) -> impl IntoViewwhere
    E: IntoView,
    F: Fn() -> E + 'static,
    P: Display,
    S: Fn() -> Pin<Box<dyn Future<Output = StaticParamsMap>>> + 'static,

Describes a portion of the nested layout of the app, specifying the route it should match, the element it should display, and data that should be loaded alongside the route.

Required Props

path: [P]

The path fragment that this route should match. This can be static (users), include a parameter (:id) or an optional parameter (:id?), or match a wildcard (user/*any).

view: [F]

The view that should be shown when this route is matched. This can be any function that returns a type that implements IntoView (like || view! { <p>"Show this"</p> }) or || view! { <MyComponent/> } or even, for a component with no props, MyComponent).

Optional Props

static_params: [S]

Creates a map of the params that should be built for a particular route.

mode: StaticMode

The static route mode

data: impl Into<Loader>

A data-loading function that will be called when the route is matched. Its results can be accessed with use_route_data.

children: Children

children may be empty or include nested routes.