Function leptos_router::create_query_signal

pub fn create_query_signal<T>(
    key: impl Into<Oco<'static, str>>
) -> (Memo<Option<T>>, SignalSetter<Option<T>>)where
    T: FromStr + ToString + PartialEq,

Constructs a signal synchronized with a specific URL query parameter.

The function creates a bidirectional sync mechanism between the state encapsulated in a signal and a URL query parameter. This means that any change to the state will update the URL, and vice versa, making the function especially useful for maintaining state consistency across page reloads.

The key argument is the unique identifier for the query parameter to be synced with the state. It is important to note that only one state can be tied to a specific key at any given time.

The function operates with types that can be parsed from and formatted into strings, denoted by T. If the parsing fails for any reason, the function treats the value as None. The URL parameter can be cleared by setting the signal to None.

use leptos::*;
use leptos_router::*;

#[component]
pub fn SimpleQueryCounter() -> impl IntoView {
    let (count, set_count) = create_query_signal::<i32>("count");
    let clear = move |_| set_count.set(None);
    let decrement =
        move |_| set_count.set(Some(count.get().unwrap_or(0) - 1));
    let increment =
        move |_| set_count.set(Some(count.get().unwrap_or(0) + 1));

    view! {
        <div>
            <button on:click=clear>"Clear"</button>
            <button on:click=decrement>"-1"</button>
            <span>"Value: " {move || count.get().unwrap_or(0)} "!"</span>
            <button on:click=increment>"+1"</button>
        </div>
    }
}