Struct leptos_router::FormProps

pub struct FormProps<A>where
    A: ToHref + 'static,{
    pub method: Option<&'static str>,
    pub action: A,
    pub enctype: Option<String>,
    pub version: Option<RwSignal<usize>>,
    pub error: Option<RwSignal<Option<Box<dyn Error>>>>,
    pub on_form_data: Option<Rc<dyn Fn(&FormData)>>,
    pub class: Option<AttributeValue>,
    pub on_response: Option<Rc<dyn Fn(&Response)>>,
    pub node_ref: Option<NodeRef<Form>>,
    pub attributes: Option<MaybeSignal<AdditionalAttributes>>,
    pub children: Children,
}

Props for the Form component.

An HTML form progressively enhanced to use client-side routing.

Required Props

cx: Scope

action: [A]

action is the URL that processes the form submission. Takes a String, &str, or a reactive function that returns a String.

children: Children

Component children; should include the HTML of the form elements.

Optional Props

method: [&'static str]

method is the HTTP method to submit the form with (get or post).

enctype: String

enctype is the MIME type of the form submission if method is post.

version: RwSignal<usize>

A signal that will be incremented whenever the form is submitted with post. This can useful for reactively updating a Resource or another signal whenever the form has been submitted.

error: RwSignal<Option<Box<dyn Error>>>

A signal that will be set if the form submission ends in an error.

on_form_data: [OnFormData]

A callback will be called with the FormData when the form is submitted.

class: impl Into<AttributeValue>

Sets the class attribute on the underlying <form> tag, making it easier to style.

on_response: [OnResponse]

A callback will be called with the Response the server sends in response to a form submission.

node_ref: NodeRef<html::Form>

A NodeRef in which the <form> element should be stored.

attributes: impl Into<MaybeSignal<AdditionalAttributes>>

Arbitrary attributes to add to the <form>

Fields
method: Option<&'static str>

method is the HTTP method to submit the form with (get or post).

action: A

action is the URL that processes the form submission. Takes a String, &str, or a reactive function that returns a String.

enctype: Option<String>

enctype is the MIME type of the form submission if method is post.

version: Option<RwSignal<usize>>

A signal that will be incremented whenever the form is submitted with post. This can useful for reactively updating a Resource or another signal whenever the form has been submitted.

error: Option<RwSignal<Option<Box<dyn Error>>>>

A signal that will be set if the form submission ends in an error.

on_form_data: Option<Rc<dyn Fn(&FormData)>>

A callback will be called with the FormData when the form is submitted.

class: Option<AttributeValue>

Sets the class attribute on the underlying <form> tag, making it easier to style.

on_response: Option<Rc<dyn Fn(&Response)>>

A callback will be called with the Response the server sends in response to a form submission.

node_ref: Option<NodeRef<Form>>

A NodeRef in which the <form> element should be stored.

attributes: Option<MaybeSignal<AdditionalAttributes>>

Arbitrary attributes to add to the <form>

children: Children

Component children; should include the HTML of the form elements.

Implementations

impl<A> FormProps<A>where A: ToHref + 'static,

pub fn builder( ) -> FormPropsBuilder<A, ((), (), (), (), (), (), (), (), (), (), ())>

Create a builder for building FormProps. On the builder, call .method(...)(optional), .action(...), .enctype(...)(optional), .version(...)(optional), .error(...)(optional), .on_form_data(...)(optional), .class(...)(optional), .on_response(...)(optional), .node_ref(...)(optional), .attributes(...)(optional), .children(...) to set the values of the fields. Finally, call .build() to create the instance of FormProps.