Trait leptos_router::History

pub trait History {
    // Required methods
    fn location(&self) -> ReadSignal<LocationChange>;
    fn navigate(&self, loc: &LocationChange);
}

The Router relies on a RouterIntegrationContext, which tells the router how to find things like the current URL, and how to navigate to a new page. The History trait can be implemented on any type to provide this information.

Required Methods

fn location(&self) -> ReadSignal<LocationChange>

A signal that updates whenever the current location changes.

fn navigate(&self, loc: &LocationChange)

Called to navigate to a new location.